requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1 # version 2.1.1
click==8.1.3
crispy-bootstrap5==0.7
crispy-tailwind==0.5.0
cryptography==38.0.4
cssselect==1.2.0
cssselect2==0.7.0
defusedxml==0.7.1
deprecation==2.1.0
diff-match-patch==20200713
Django==4.1.5
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-import-export==3.0.1
django-jsignature==0.11
django-localflavor==3.1
django-non-dark-admin==2.0.2
django-qr-code==3.1.1
django-request-logging==0.7.5
django-tenants==3.4.7
django-tinymce==3.5.0
et-xmlfile==1.1.0
fonttools==4.38.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
localflavor==1.9
lxml==4.9.1
MarkupPy==1.14
mollie-api-python==2.13.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
packaging==21.3
pikepdf==6.2.4
Pillow==9.3.0
prettyjson==1.0.3
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.2
pydyf==0.5.0
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
pyparsing==3.0.9
PyPDF2==2.11.2
PyPDF3==1.0.6
pyphen==0.13.2
pyquery==1.4.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-stdnum==1.18
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
segno==1.5.2
six==1.16.0
sqlparse==0.4.3
svglib==1.4.1
tablib==3.2.1
tinycss2==1.2.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.13
webencodings==0.5.1
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.2
